feat(lanes): split lanes.e2e.ts for parallel execution
